Fix tracking of corrections, add deletion costs
There was an error before where returning to an instruction after
correction instructions was not correctly identified, as the remaining
blocks of a high level object were part of the identity of the object
to be instructed.  This commit fixes that.  In addition, we do not
consider block instructions for intermediate corrections (as a block
cannot be partially built).

The code now can also use the number of block deletions as a cost
metric instead of the completion times.

bump to java 16 (records!)